requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
django-filter
django-countries
requests
Pillow